// backend/src/modules/reviews/review.controller.d.ts
import type { RequestHandler } from "express";
declare class ReviewController {
    listForProduct: RequestHandler;
    listAll: RequestHandler;
    getById: RequestHandler;
    listMine: RequestHandler;
    create: RequestHandler;
    updateContent: RequestHandler;
    setStatus: RequestHandler;
    delete: RequestHandler;
}
export declare const reviewController: ReviewController;
export {};
//# sourceMappingURL=review.controller.d.ts.map